# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Travel Risk API",
    "description": "Backend of the tourism travel safety risk assessment platform",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/regions": {
      "get": {
        "tags": [
          "regions"
        ],
        "summary": "List Regions",
        "operationId": "list_regions_api_regions_get",
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Regions Api Regions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "regions"
        ],
        "summary": "Create Region",
        "operationId": "create_region_api_regions_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegionCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/regions/{region_id}": {
      "get": {
        "tags": [
          "regions"
        ],
        "summary": "Get Region",
        "operationId": "get_region_api_regions__region_id__get",
        "parameters": [
          {
            "name": "region_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Region Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "regions"
        ],
        "summary": "Patch Region",
        "operationId": "patch_region_api_regions__region_id__patch",
        "parameters": [
          {
            "name": "region_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Region Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegionPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "regions"
        ],
        "summary": "Delete Region",
        "operationId": "delete_region_api_regions__region_id__delete",
        "parameters": [
          {
            "name": "region_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Region Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/regions/{region_id}/respondents": {
      "get": {
        "tags": [
          "respondents"
        ],
        "summary": "List Respondents",
        "operationId": "list_respondents_api_regions__region_id__respondents_get",
        "parameters": [
          {
            "name": "region_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Region Id"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Year"
            }
          },
          {
            "name": "month",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Month"
            }
          },
          {
            "name": "accommodation",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Accommodation"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Respondents Api Regions  Region Id  Respondents Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "respondents"
        ],
        "summary": "Create Respondent",
        "operationId": "create_respondent_api_regions__region_id__respondents_post",
        "parameters": [
          {
            "name": "region_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Region Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RespondentCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespondentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/respondents/{respondent_id}": {
      "patch": {
        "tags": [
          "respondents"
        ],
        "summary": "Patch Respondent",
        "operationId": "patch_respondent_api_respondents__respondent_id__patch",
        "parameters": [
          {
            "name": "respondent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Respondent Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RespondentPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespondentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "respondents"
        ],
        "summary": "Delete Respondent",
        "operationId": "delete_respondent_api_respondents__respondent_id__delete",
        "parameters": [
          {
            "name": "respondent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Respondent Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LatestResult": {
        "properties": {
          "mu": {
            "type": "number",
            "title": "Mu"
          },
          "risk_class": {
            "type": "string",
            "title": "Risk Class"
          },
          "evaluated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Evaluated At"
          }
        },
        "type": "object",
        "required": [
          "mu",
          "risk_class",
          "evaluated_at"
        ],
        "title": "LatestResult"
      },
      "RegionCreate": {
        "properties": {
          "name_uk": {
            "type": "string",
            "minLength": 1,
            "title": "Name Uk"
          },
          "name_en": {
            "type": "string",
            "minLength": 1,
            "title": "Name En"
          },
          "xi": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Xi"
          },
          "delta_level": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "D1",
                  "D2",
                  "D3",
                  "D4",
                  "D5"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Delta Level"
          }
        },
        "type": "object",
        "required": [
          "name_uk",
          "name_en"
        ],
        "title": "RegionCreate"
      },
      "RegionOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name_uk": {
            "type": "string",
            "title": "Name Uk"
          },
          "name_en": {
            "type": "string",
            "title": "Name En"
          },
          "xi": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Xi"
          },
          "delta_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delta Level"
          },
          "respondent_count": {
            "type": "integer",
            "title": "Respondent Count"
          },
          "latest_result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LatestResult"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "name_uk",
          "name_en",
          "xi",
          "delta_level",
          "respondent_count",
          "latest_result"
        ],
        "title": "RegionOut"
      },
      "RegionPatch": {
        "properties": {
          "name_uk": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name Uk"
          },
          "name_en": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name En"
          },
          "xi": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Xi"
          },
          "delta_level": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "D1",
                  "D2",
                  "D3",
                  "D4",
                  "D5"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Delta Level"
          }
        },
        "type": "object",
        "title": "RegionPatch"
      },
      "RespondentCreate": {
        "properties": {
          "ext_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ext Id"
          },
          "year": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          },
          "month": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Month"
          },
          "accommodation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accommodation"
          },
          "ratings": {
            "additionalProperties": {
              "type": "integer",
              "maximum": 5.0,
              "minimum": 1.0
            },
            "type": "object",
            "title": "Ratings"
          }
        },
        "type": "object",
        "required": [
          "ratings"
        ],
        "title": "RespondentCreate"
      },
      "RespondentOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "ext_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ext Id"
          },
          "year": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          },
          "month": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Month"
          },
          "accommodation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accommodation"
          },
          "ratings": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Ratings"
          }
        },
        "type": "object",
        "required": [
          "id",
          "ext_id",
          "year",
          "month",
          "accommodation",
          "ratings"
        ],
        "title": "RespondentOut"
      },
      "RespondentPatch": {
        "properties": {
          "ext_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ext Id"
          },
          "year": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          },
          "month": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Month"
          },
          "accommodation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accommodation"
          },
          "ratings": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "integer",
                  "maximum": 5.0,
                  "minimum": 1.0
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ratings"
          }
        },
        "type": "object",
        "title": "RespondentPatch"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}